requirements.txt
beanie 
datasets 
fastapi 
passlib[bcrypt]
fastapi_mail 
huggingface_hub 
langchain 
langchain_community 
langchain_text_splitters 
matplotlib 
motor 
openai 
pandas 
plotly 
pydantic 
groq
pydantic_settings 
python-dotenv 
ragas 
redis[hiredis]
tqdm 
uvicorn 
PyJWT
Jinja2
python-multipart
pacmap 
weaviate-client
websockets
minio
langchain-huggingface
pymupdf